Refactor directory writing a bit
Summary:
Extract the directory writing logic into a helper function that just
writes a content map to a directory. Why do this? Because it's useful for the
diffs (so useful that until this commit, I was just creating a silly
MemoryBackedTypeshed so that I could get this behavior).

This allows us to no longer pretend the diffs are "a typeshed".

Reviewed By: kinto0

Differential Revision: D51450410

fbshipit-source-id: f99e15f9742a1dc873b7bc37ce69098d6646f9b6