Batch type errors on making request
Summary:
**Resubmission of D50901486**, since it's been sitting for 2 weeks and some changes needed to be made to rebase it.

We would like to be able to batch type error requests for files, instead of needing to send separate requests for each file.

This diff makes the changes required for the new interface on the client side. This involves leaving some functionality for parsing the old version of the request (just the name), since at some point we want to migrate back to that name for performing type error requests.

Reviewed By: kinto0

Differential Revision: D51404224

fbshipit-source-id: ff98a1dd9a74556d0705e0685f16bb6a25f2127c