ignore build failures on code navigation
Summary:
87% of uncaught exceptions (1.2k today) in shadow mode are from buck crashing. this is a huge reason the server is down.

This diff copies the strategy described here (D48591009) in the code navigation server to ignore build failures and defer updates until a successful build.

This diff does not change the user messaging. It will appear as "Building" in cases where buck is unable to build and artifacts may be stale. This is OK since no buck-based repositories use pyre for code navigation yet. Before the release, I think the UI should be improved as well. For example, a "build results stale" message should show up instead of "pyre ready".

Reviewed By: jasontatton, grievejia

Differential Revision: D51434368

fbshipit-source-id: 0ddb6144e8dd79bc180ac0ed9af4f9c5b65b539d