Tweak end-to-end patch interface + implementation
Summary:
Following up on diff review of D51161396, I have two related
changes here:

- change `source` and `target` options to be `source_root` and
  `target_root`
- lift the overwrite flag handling up to the top level, which makes the
  lower-level logic clearer.

I'm a little on the fence about the second change; I like to have
my click handlers do nothing at all so that you can execute *exactly* the
same logic from Python, and that's not generally possible if we put logic
directly into the handler.

That trades off against the desire to force all python interfaces to have
expressive type signatures so that, for example, the `overwrite` and
`diffs_directory` arguments have no coupling between them (it's not really
practical to make clis, which are string-only, that expressive).

But I guess there's no precedent for trying to make the click handlers
accessible from python in our code anyway, so that's just a silly thing I
sometimes try to do (in a big codebase it's not silly, since you can avoid
startup overhead this way. Some tools would be vastly improved by doing this.
But our code is pretty quick to start up.)

Reviewed By: kinto0

Differential Revision: D51450407

fbshipit-source-id: b6e96bd315052128afaf54210742a592a137f8a5