`s/target/destination/` in typeshed patcher
Summary:
Kyle pointed out in a review of an earlier diff that using
"source / target" (which we often do because it's a standard naming
convention in linux tools) can be confusing in the context of buck-based
builds.

Replacing "target" with "destination" makes things clearer.

This diff replaces all uses of "target" as an output path with
"destination".

There are still `target` variables in the transforms module,
but that's unavoidable; there's a Python AST node named target (the LHS of
an assignment is a "target pattern").

Reviewed By: kinto0

Differential Revision: D51471457

fbshipit-source-id: a1a4003793e7f9b4c1a3b7a48ada9fdc239ff06e