Make forward-compatible changes to container operations
Summary:
The next version of the Jane Street `Core` library no longer exposes
a number of functions in the output modules of container `Make`
functors, instead preferring to use the polymorphic functions directly
available in `Core.Map`, `Core.Hashtbl`, etc.

This diff flips to that interface in preparation for a version bump.

Reviewed By: grievejia

Differential Revision: D51533931

fbshipit-source-id: b2278e2424c3de53f98ef06341445d4094dc0127