Attach breadcrumb `propagated-return-sink` for taint that originates from the return statement
Summary:
So that Pysa still detects flows that contain sink taint that originates and is
propagated from the return statement, but meanwhile we allow only viewing
traces that don't contain such sink taint. This can be achieved by filtering
out traces that contain breadcrumb `propagated-return-sink`.

Breadcrumb `propagated-return-sink` is intentionally only added to the backward
analysis, but not the forward analysis. The backward analysis propagates the
sink up, which might be surprising to the users and hence adding the breadcrumb
helps to explain the unexpected behavior. By contrast, the forward analysis
simply detects flows at the return statement. It is uncontroversial that this
is the desired behavior.

Name `propagated-return-sink` reflects the fact that this is about sinks that
originate from the return statements AND is propaagated up (i.e., not all sinks
that originate from the return statements have this breadcrumb). An alternative
is `ReturnSink`, but this is confusing because it might suggest a callable
"return"s a sink, which does not make much sense.

Reviewed By: arthaud

Differential Revision: D51510212

fbshipit-source-id: 7fb1e1e176477c5d492e42df3be4a6724a94d932