Tito for captured variables test
Summary:
Let's demonstrate what tito for captured variables would look like.

Surprisngly enough, the generated model is correct (check the generated expected file). This is due to the fact that the previous work for nonlocal captured variable reads added `CapturedVariable` as an AccessPath that flowed into the arguments. Returning that variable completes the taint in taint out path, so we get models for free.

Note, the models aren't being used correctly, as demonstrated with no issue being found.

Reviewed By: arthaud

Differential Revision: D51541916

fbshipit-source-id: b5787d262d1a42452d0b14c844ed2ac6524b8a8b