Rename Type.{instantiate -> apply_mapping}
Summary:
This function just renames parts of a potentially-complex type
recursively. It *can* be used as part of instantiating type variables,
but the majority of use cases have nothing to do with
instantiation.

Rename it. Also rename `constraints` (whose name collides with the term
we are using in this codebase for type variable bounds) and add
a doc comment since this is one of the most central utility functions
in Type.ml

Reviewed By: connernilsen

Differential Revision: D51592667

fbshipit-source-id: 9138835c74238abbedc92f71b38ed567c2f1f102