Rename all_attributes [1]: instantiated_attributes
Summary:
We have various methods and functions named `all_attributes`, where
sometimes functions with different names are actually the same and sometimes
functions with the same name are actually different. It's all very confusing!

This is the first half of a refactor to always use consistent names:
- functions / methods returning a list of instantiated attributes shall
  be named `instantiated_attributes`
- functions / methods returning a list of uninstantiated attributes shall
  be named `uninstantiated_attributes`

Keeping the names straight makes code search easier, and is a step toward
building a decent mental model of what the code is doing.

Reviewed By: kinto0

Differential Revision: D51644482

fbshipit-source-id: 543cbe5f2e1280424df2b40e1bf6a139f706367a