Remove dangling TODO
Summary:
We closed T41143153 (to investigate making module types
more first-class) with the message:
> Don't think it's possible to support this in the foreseeable future -- this requires module tracking to depend on type checking, which doesn't interact well with our current infra.

We probably shouldn't have a TODO dangling in our code pointing to a task we've
marked as probably infeasible without a major overhaul (not that we shouldn't
consider such an overhaul someday, but a TODO on an existing function isn't
going to be useful for that anyway).

Realistically anything like this would probably be related to qualification.

Reviewed By: grievejia

Differential Revision: D51642747

fbshipit-source-id: cb355e019d38cc2ed24698fb98ceb5761eb892cb