Remove GlobalResolution.resolved_type
Summary:
The GlobalResolution module is a wrapper for queries that hit
environment cache tables. We shouldn't have any functions in there
that are just pure logic on individual bits of data, because this
obscures the fact that such functions are simple pure operations.

It's particularly unnecessary given that there's only one use case
anyway, in `pyre infer` code. Let's just use the underlying pure
logic directly.

Reviewed By: grievejia

Differential Revision: D51642635

fbshipit-source-id: e27c68d05a6678c59f21203f054e1c6882579f0b