remove circular dependency in daemon_query_failer
Summary:
daemon_query_failer relies on `DaemonQueryFailure` which conceptually belongs to the daemon_querier. we should decouple these and provide a separate class to return in cases of DaemonQueryFailer's failure since the daemon_querier uses these failures.

the `DaemonConnectionFailure` also could be implemented in this file as well at some point in the future but it isn't as necessary.

Reviewed By: inseokhwang

Differential Revision: D51681306

fbshipit-source-id: f73139c0e355c8199d4ca95a8405f373245af378