Move instantiated_annotation into a module
Summary:
The rationale for this will be clearer in the next diff: in order
to start decoupling the logic of queries from the environment stack
I need to start moving types out of AttributeResolution, and the definition
of *uninstantiated* attributes needs to move.

Using modules for both makes things more consistent and easier to skim.

Reviewed By: arthaud

Differential Revision: D51735294

fbshipit-source-id: 60303542ac76c64733c6ff94151a53636703cdaf