Move UninstantiatedAttribute to AnnotatedAttribute
Summary:
This type doesn't belong inlined into AttributeResolution, which
is part of the environment stack. Instead, we want types to live
in pure-logic modules; in this case the type is always used inside
of an `AnnotatedAttribute.t`, so it makes sense to move the type
definition there.

The resulting code is pretty messy, I'm very confident it could be
reorganized to make the type names more succinct and the type (and
data constructor) names better grouped together. But even though
it's ugly this is a step forward in terms of modularity.

Reviewed By: grievejia

Differential Revision: D51735293

fbshipit-source-id: 4d51c058ae19bb4a2bc47dce6f9d2ab96f998596