Send type check request on LSP startup
Summary:
When the Global Lazy type checker starts up, there's a bug where type errors won't be returned for any files that were opened before startup is complete. To get type errors for any already-opened files, the user needs to edit a (any) file and save it, and only after that will type errors be populated.

This occurs because of a race condition where (see test below for an example):
- VSCode sends `didOpen` (and other notifications) as soon as the language server starts up
- The language server handles forwarding notifications to the daemon as its starting up/before the queriers are registered
- The notification requests to the daemon error since they are trying to send before the queriers/daemon are ready
- The `didOpen` requests are resent at the end of querier startup, and all other notification types are dropped

To solve this, I've added an `asyncio.Event`, which allows querier requests to wait until `client_setup()` is called to register queriers.

Reviewed By: inseokhwang

Differential Revision: D51594826

fbshipit-source-id: dac0697e7c700da6b17984c3a3b54b8fd91b5946