Add basic support for explicit any annotations to pyre statistics run
Summary:
This has been a request from instagram to have pointers not only to any expressions but to explicit any annotations. This is a basic implementation that is in pyre statistics. Benefits of doing it here allow us to easily plug the data into the existing pipelines.

I'm aware that this likely could have been done in the ocaml side - but for now wanted to keep it consistent with the rest of statistics

Reviewed By: stroxler

Differential Revision: D51833335

fbshipit-source-id: 87961de71a99b9b71ff6647eab69d7bde0bc9da9