Combine `self_taint` and `callee_taint` into `implicit_argument_taint`
Summary:
The hidden invariant is that we never expect `self_taint` and `callee_taint` to
both exist, because both refer to the taint on the implicit argument at a call
site (regardless whether the implicit argument is `self` or the entire callee).
This is now explicated by this diff.

The code is also simplified, since we no longer need to pass around both
`self_taint` and `callee_taint`. We also don't need to handle the unexpected
cases where both may exist (because both are currently optional typed).

We can do the same for the forward analysis in the next diff.

Reviewed By: arthaud

Differential Revision: D51770876

fbshipit-source-id: 0e8b1da7bc0e8ff892cea35fc5f9222a56ca1aa2