Rename some locals in set_class_summaries
Summary:
Calling the class name "primitive" and the set of classes in the current module
"new annotations" was confusing to me. I can see how these names make a kind of
sense (becasue we're using the pun that Identifier.t = Type.Primitive.t, and
we're using "annotation" as a synonym for "type" which is then a
not-really-but-almost synonym for "class"), but calling things what they
actually are makes it much easier to follow.

Reviewed By: grievejia

Differential Revision: D51831841

fbshipit-source-id: 6cc6bcd872ec4b98a11a368bf570d4b82098ccee