Separate computation in UGE [7]: [easy] inline ResolveExportItem
Summary:
I'm beginning to look at splitting out the pure-computation
aspects of UGE that are downstream of the module component tables, and
one of the first things I got confused by was `ResolveExportItem`; I
breifly thought there was maybe another type defined in UGE that needed
to be extracted.

It turns out this module is an implementation detail of `resolve_exports`.
For now I'm inlining it there; this makes the function bigger, which I'm not
crazy about, but it also makes the organization of the code as a whole clearer
which is a higher priority for me right now.

Reviewed By: grievejia

Differential Revision: D51857612

fbshipit-source-id: 7d06da8d319e37a4414f8a5694d01a8d0f0c2b8e