Queries for AttributeResolution [2/2]
Summary:
This commit completes the in-place (by that I mean source-level but not
module-level) separation of AttributeResolution business logic from the
environment stack, by making the Queries object big enough to encompass all
reads from previous environment layers.

The `base` class constructor now just takes a queries record as input, and no
longer ever passes around `?dependency` manually (which, by the way, was a
disaster waiting to happen - the dependency-tracking is *never* type safe
because you could leave it off and the compiler wouldn't warn you!), instead
effectively currying it in the creation of the record.

Reviewed By: grievejia

Differential Revision: D51768048

fbshipit-source-id: 3191c7b2b0a71ab2c53897597dbf37f00bd90ceb