Separate computation in UGE [6]: Finish extracting incoming data logic
Summary:
This commit completes the IncomingDataComputation extraction for UGE:
the fallback to an implicit when no source is found but a module is tracked
is now handled in terms of an abstract Queries interface and the general
`module_components` that result instead of being directly hard-coded into
the control flow.

The UGE storage layer remains very complex because lazy loading and dependency
tracking are both quite complicated, but the underlying analysis logic is actually
quite simple and fits in the first 200 lines!

Next up will be outgoing data computation, which is mostly already done
(unlike other environment layers, the UGE ReadOnly is already *almost* the way
we want things to work after splitting storage and compute - the only thing we have
to deal with is the direct representation of `AstEnvironment`.

Reviewed By: grievejia

Differential Revision: D51831838

fbshipit-source-id: 4f038f16e4561b1523e769ca175d03960c1623b1