Queries for AttributeResolution [1/2]
Summary:
As of this diff I'm going to begin removing direct use of
environments from the AttributeResolution analysis logic.

AttributeResolution is so big that I'm not going to do this
all in one diff; instead I will move over a couple use cases
at a time to keep diffs a little smaller.

This commit kicks off the process by moving everything *above*
the `base` class definition to use a queries object where the
environment and dependency key are curried.

Reviewed By: grievejia

Differential Revision: D51768049

fbshipit-source-id: dee34b9748dc61a6271ef8dc60a5631369e2e17e