Separate computation in UGE [4]: [easy] Move set_module_data up
Summary:
The next stage of my refactoring is to fully separate the extraction
of module components (function definitions, class summaries, etc) from
the storage layer.

This is easier if we first move `set_module_data` to live next to all
the other setter functions; it's a little confusing that this function
isn't already next to the other functions.

Reviewed By: grievejia

Differential Revision: D51831837

fbshipit-source-id: 54d6ed542267c9ded1d6579cc1b810afe4d0ad44