Separate computation in UGE [5]: Refactor component extraction
Summary:
I'm working on cleanly separating the logic for extracting module components
like class summaries and function definitions from storage, and my goal is to
have a function that creates a single record of components from a source as
part of the abstract interface, with all of the storage handling in a separate
place.

At the moment the two are still a little bit intertwined for two reasons:
- the control flow has baked into it a loop of "compute this thing, store it,
	repeat"
- we also have a try/catch to deal with implicit modules

This diff addresses the former concern by extracting a record representing
all the components we extract from the source (module metadata, class summaries,
unannotated globals, and function definitions) into a record, so that we can
run a single pure computation per module load.

In the process I've renamed the `IncomingDataComputation` module to `ModuleComponents`
because it's actually just pure functions with no query interface needed.

I'll address the second issue shortly, by making a proper `IncomingDataComputation`
that will depend on an abstract API for `AstEnvironment` and `ModuleTracker`.

Reviewed By: grievejia

Differential Revision: D51831839

fbshipit-source-id: 2917e0eb858e28ba0003aea7aa862a3807bcfdac