Separate computation in UGE [3]: distinguish names of things
Summary:
At this point we've successfully separated out the low-level compute
of extracting things like class summaries from a source, so that that
low-level logic is now independent of shared memory.

But the overall logic is still coupled right now; I want to go one
step further and have a pure function to get everything from a source
and put it in a single record, which the storage layer can then process
as it wants.

I ran into a pretty silly blocker on this: the fact that the table names
and the things inside the table want to be the same name is irritating;
for example I want to put `class_summaries`, the result of extracting
class summaries from source, into the class summary table but that table
itself is named `class_summaries`.

I could deal with this using local renames, but the truth is it's easier
to skim code when there aren't name collisions so I decided to just make
the rename everywhere: this diff renames all the shared memory tables
to be of the form `{name_of_value}_table`.

This is entirely private to the UGE implementation - none of these names are
exposed in the interface.

Reviewed By: grievejia

Differential Revision: D51831843

fbshipit-source-id: adfd404bc72d31a25493895fa17c02c0b3e48c3e