Combine `self_taint` and `callee_taint` into `implicit_argument_taint` (Forward analysis)
Summary:
This is the same as D51770876, but for the forward analysis. We pass in
`implicit_argument_taint` to `apply_call_target`, instead of both `self_taint`
and `callee_taint`, so that the intention is clear -- `self_taint` and
`callee_taint` are only used as implicit argument taint, and only one of them
should exist.

Reviewed By: arthaud

Differential Revision: D51776200

fbshipit-source-id: 4d95c7d86147bde5d81f0768b29b2887e1a917f1