Set `implicit_receiver` based on `is_static_method` and `is_class_method`
Summary:
At a call site:
- `is_static_method` should always imply `implicit_receiver` is false.
- `is_class_method` should always imply `implicit_receiver` is true.

This is not a part of the previous diff D51234140, because D51234140 adds new
information about whether a callee is a static method, which can now be used in
this diff.

This diff also adds an implicit argument for implicit calls to `__new__`,
because `__new__` are static methods and static methods don't have implicit
receiver. However, according to the python doc, `__new__` takes the class as
the first argument. Hence we need to add this implicit argument at the call sites.

Reviewed By: arthaud

Differential Revision: D51689363

fbshipit-source-id: 831f92410482c8030dcd8855001dc61653467ffc