show better hover results for boundmethod
Summary:
BoundMethods do not look pretty in hover results:
{F1167042631}

Not to be confused with [python's bound methods removed in python 3](https://www.bhavaniravi.com/python/advanced-python/bound-vs-unbound-functions-in-python), pyre has an implementation detail known as bound methods for some typechecking analyses. The difference is any method on an object `<>.foo()` vs an alias itself `foo()`.

In hover results, users do not care about this implementation detail. this diff uses the same formatting as normal function calls for boundmethods to print them nicer for general users.

Note: doing this includes the self parameter even though callers do not need to worry about that argument. it is a different color but if users complain we can remove it
Note: during pyre typechecking, this `pp_concise` is not used and `pp` is used. in type errors, we will still show the same behavior.

Reviewed By: grievejia

Differential Revision: D51898082

fbshipit-source-id: 9cbefb392a88b20d4ed40f315b29357146ac85fc