Turn off dependency tracking more fully
Summary:
I realized when trying to minimize the dependency table for pysa and
some other use cases that we actually don't honor the `track_dependencies`
flag in many of our environment layers. The managed cache layers do
honor it, by bypassing the shared-memory tables entirely, but all the
other layers are still passing dependencies!

This comes with more than one cost; it forces us to allocate a big table
(which is why I'm trying to turn it off), and also introduces aggregation
overhead in map-reduce operations.

This diff tries to turn off the dependency tracking itself, which I hope
will come with a (small) perf boost for full check and pysa. The next diff
will try to make the table tiny when we aren't tracking dependencies, to
reduce our memory footprint and get another perf win.

Reviewed By: alexkassil

Differential Revision: D51908475

fbshipit-source-id: adddc85a3c81f897934710bb53bdd4ca429826de