Group arguments of `apply_call` that are only related with class intervals
Summary:
So that the readability is improved. Currently `apply_call` takes 13 arguments,
which is difficult to understand. This diff reduces to 10 arguments.

Maybe we can further group `is_class_method` and `is_static_method` as well,
since they are also used only when applying the class intervals. I decided
against it for now, because:
- the other four arguments can be conveniently grouped / expressed as
  `CallInfoIntervals.t`, and
- when calling `apply_call`, argument `call_info_intervals` is almost always
  stored "directly" (i.e., without any modifications) into the taint, except
  that sometimes field `caller_interval` need to be updated based on the
  results of applying class intervals -- this is also made explicit in this
  diff with the OCaml record update.

Reviewed By: alexkassil

Differential Revision: D51910033

fbshipit-source-id: 1a47c5fa85a04baedf99aa67b1ea1283a56913e7