Detect calls to `cls` in addition to calls to `self`
Summary:
# Context

A key idea in applying class intervals is that, on any call chain (i.e., a
sequence of method calls), we should only apply class intervals onto a
continuous subsequence of calls in which every call is on the *same* object.
Consider the following example:
```
a = A()
a.foo()  # Reset the class interval to top

def A:
  def foo(self):
    self.bar_1()  # Apply class interval of A
  def bar_1(self):
    self.bar_2()  # Apply class interval of A
  def bar_2(self):
    b = B()
    b.baz()  # Reset the class interval to top
class B:
  def baz(self):
    pass
```
Now, we can only apply the class intervals at the call sites to self.bar_1 and
self.bar_2 (but not other call sites), because this is a sequence of calls onto
the same object `self`.

# This diff
Previously, we only detected calls to `self`, which is used to tell if there
exists a sequence of calls to `self`. This is one case of a continuous sequence
of calls to the same object. Another case is a continuous sequence of calls to
`cls` when calling static methods. This diff further detects calls to `cls`.

Reviewed By: alexkassil

Differential Revision: D51871818

fbshipit-source-id: e5ce2c755eb592aa24aad324dda29200466981c0