Add dune for zstd
Summary:
Adding a dune file for ZSTD.
The previous diff was a mere copy of ZSTD over.
This diff adds a dune file and thus plugs ZSTD into PYSA.
We only need compression/decompression and common parts of the library, not dictionary training, so the dune hooks only those parts in.

Reviewed By: arthaud

Differential Revision: D51444344

fbshipit-source-id: c8f3803044896aa3fc6b00ca509ed8560cb0b2ac