use GlobalResolution where we can in query
Summary:
We shouldn't be manually unpacking a UGE from GlobalResolution
in order to call a function that's already defined in the GlobalResolution
api.

Note that this is a tiny code cleanliness win on its own, but it's
really coming from me starting to audit use cases of GlobalResolution
with an eye toward refactors to support per-target-typechecking.

Reviewed By: grievejia

Differential Revision: D51964997

fbshipit-source-id: 2b16038977f69d84553cddec13949737c7b43ce2