Store class intervals as keys in the map that represents the taint information
Summary:
# Context
We currently merge different taint kinds that are associated with different
intervals, which might cause false positives.

# This diff
This diff solves the above problem, by moving the intervals outside
`LocalTaint` so that it is now along side `CallInfo`, which is the key in the
map of `ForwardTaint` and `BackwardTaint`.

The changes in the *.models files are due to
1. "splitting" the taint based on the class intervals. Before this diff, they
are merged together because the interval information is stored in the value
part of the map.
2. the changes in `CallInfo.to_json`, which now emit intervals. Before this
diff, the intervals are emitted by json-ify `LocalTaint`. This causes the
intervals to be moved around in the *models files.

# More context
The meaning of a key-value pair `((CallInfo, CallInfoIntervals), LocalTaint)`
in `ForwardTaint` or `BackwardTaint` is that, taint `LocalTaint` is generated
at call site `CallInfo` (say inside callable `c`). Meanwhile, this `LocalTaint`
can be further propagated to the caller of `c`, only if the caller class
overlaps with interval `CallInfoIntervals.caller_interval`.

Reviewed By: arthaud

Differential Revision: D50950967

fbshipit-source-id: cb8d0fdbd8dd479585a50604d2f4e767d49532ec