log exceptions in asyncio.create_task
Summary:
## Problem
When an async request raises an exception, we don't crash. This is because we `asyncio.create_task` the function sending it to event loop for further evaluation. Instead, we get logs like this on stderr:
```
2023-12-08 11:52:53,086 [PID 1338478] ERROR Task exception was never retrieved
future: <Task finished name='Task-28' coro=<PyreLanguageServerDispatcher.dispatch_nonblocking_request() done, defined at /data/users/kylei/fbsource/buck-out/v2/gen/fbcode/ce2fc14956a81fb9/tools/pyre/facebook/client/__pyre-debug__/pyre-debug-inplace#link-tree/tools/pyre/client/commands/pyre_language_server.py:1450> exception=Exception(')')>
```

This obviously is not ideal considering we would like to know about issues with the language server before receiving user reports.

## Solution
This diff adds telemetry events for any request with the decorator, starting with definition requests (the only one rolled out mode in fbcode shadow mode).

## Other Approaches
Something like this

```
        def done_callback(task: asyncio.Task, request: json_rpc.Request) -> None:
            exception = task.exception()
            if exception is not None:
                raise exception
            self.outstanding_tasks.discard(task)
...
request_task.add_done_callback(done_callback)
```
does not work. this is because the futures library catches the exception we re-raise. It prints out `ERROR Exception in callback PyreLanguageServerDispatcher.dispatch_request...` and does not crash the program

There is also another option that uses `add_done_callback` and logs, but it is more complicated. because we are in the LanguageServerDispatcher instead of the LanguageServer, we have less ability to log what's going on. See that implementation here: D52045543. I prefer this due to the enhanced logging capabilities.

## Future work
The language server should probably not be asynchronous. We will discuss how to improve the design of the language server offline, most likely removing this change in the process. Before that's done, though, it makes sense to put this decorator on all requests.

Reviewed By: inseokhwang

Differential Revision: D51994746

fbshipit-source-id: 02aa08d654018498d5e777e5ed92081a3df84b7b