Ignore build failures from working set updates
Summary:
D51434368 ignored errors on buck when coming from the `update_sources` api. It did not change anything for `update_working_set`, of which there are a bunch of buck failures when TARGETS is broken. These events happen on file open / close. If we open a file and there is a build failure, we should ignore it. Instead, we should resort to what we have access to without it.

This is eventually consistent since a successful buck build will update the artifact path for all files once it builds.

Users will not know that their code navigation is using outdated buck but there is already a task for that here: T170690268

Reviewed By: grievejia

Differential Revision: D52055348

fbshipit-source-id: 6509e13ca053a66296128004eb849513e4af7db0