Strip env access from GlobalResolution [victory!]
Summary:
This is what the stack has been working toward:
do not expose the underlying environments or module
tracker at all from GlobalResolution.

At this point it presents a complete (and always lazy)
interface that's needed for type checking. The implementation
remains coupled to the environment stack, but we
can relatively easily turn it into an abstract interface
in a follow-up stack.

Reviewed By: kinto0

Differential Revision: D52062173

fbshipit-source-id: cee0b24a1d6fa89a6f73a018638e0570e014676f