tweak modelQueryExecutor [1]
Summary:
The pysa code occasionally needs project-global data; the
attributes logic in `modelQueryExecutor` is an example of this.

Today it's getting that data through GlobalResolution by grabbing
an environment out of it and using `all_classes`; this is definitely
not something we want to support in the GlobalResolution interface
long-term since lazy module tracking doesn't even support this
kind of query.

So instead refactor the pysa code to take an environment as input
directly everywhere along the control flow path to this use of
`all_classes`.

NOTE: Now isn't the time, but I actually really want to create
a `PysaResolution` interface at some point to encapsulate all the
Pysa use of Pyre environment data, and in the process document
(in code) what exactly that interface is. This is something I'd
like to tackle sometime in 2024; now isn't the time since it
doesn't directly block pyre-drives-buck.

Reviewed By: arthaud

Differential Revision: D52047272

fbshipit-source-id: 6eacbf4e06fcc5a1dac9ba1f02c09855e682e3c8