Avoid some environment access through GR
Summary:
I'm working on stripping direct access to environments from the
GlobalResolution interface, which is a prerequisite to fully decoupling
computation from storage in code that relies on GlobalResolution
(namely type checking, but other code including Pysa also uses
GlobalResolution).

I don't want to do it all in one diff, so I'll be doing it in
bits and pieces; this commit makes the change for `pyre infer`
and Pysa's `fetchCallables` modules using GR to get to an AstEnvironment
in order to call `get_processed_source`. In both cases they already
have access to an `AnnotatedGlobalEnvironment`, so we can just use
that instead.

This obviously means that for now both those use cases remain coupled
to storage, but that's okay; we don't have short-term plans to migrate
either of these to buck-calls-pyre so there's no need to change anything
here for at least 6 months.

Reviewed By: connernilsen

Differential Revision: D52044964

fbshipit-source-id: ea703bc590e4d79cb7496d41eba4d10a4e80f3e8