unify transitive successor interface, rename
Summary:
In this commit I'm continuing to try to clean up GlobalResolution interfaces
and eventually eliminate direct access to environments through it; in this
case the ConstraintsSetTest was grabbing a raw environment in order to run
`is_transitive_successor`.

An issue I ran into along the way is that the `GlobalResolution` and
`ClassHierarchy` interfaces for `is_transitive_successor` were not compatible;
the variable names didn't agree! The code was accessing a raw environment
specifically to work around this problem.

I decided that the name `is_transitive_successor` wasn't very informative
anyway; instead I renamed it to `has_transitive_successor` and made the
`~successor` a named argument / the predecessor a positional argument,
so that the name actually translates reasonably to English, as in
"does <predecessor class> have a transitive successor of name ~successor"?

I still hate the naming here, since "successor" actually means "ancestor"
in the type system; we're using terminology coming from thinking about
the MRO in a context where the MRO doesn't matter that much and thinking
about the inheritance DAG would make a lot more sense. But renaming in
terms of ancestors / descendants is a much bigger change than I want to make
right now and might even be controversial. At the moment I just want
consistent interfaces.

Reviewed By: connernilsen

Differential Revision: D52044965

fbshipit-source-id: ae4f6b98f4926c26ff3838f2a2982ba8fb270f01