Clean up is_typed_dictionary
Summary:
The existing `is_typed_dictionary` implementations
on classSuccessorMetadataEnvironment and globalResolution
are not compatible:
- one takes class name as an argument, and returns None
  if the class isn't found
- the other takes a type as an argument, and returns false
  if the class isn't found

I'm not going to worry too much about the different return
types at this point, but these two functions need different
names in order for things to be simple and consistent across
the stack.

So, I've renamed the class-name-indexed version to
`is_class_typed_dictionary`, and in the process extracted this into a
proxied function call in GlobalResolution, whereas the type-indexed version
is now a logical function in GlobalResolution.

Reviewed By: grievejia

Differential Revision: D52101098

fbshipit-source-id: af1e69bde7ba2ffe0dce3bf6e5e13a5e7b5545b6