Isolate most remaining pure-logic functions
Summary:
Strip environment access from the implementation of most of
the remaining globalResolution functions that aren't just
proxies to environment layers.

This did require an API change to not muck with the interface of
`consltraints_solution_exists`. I'm not thrilled with the result - the
`get_typed_dictionary_override` argument, which isn't useful in most
places, now has to be set in typeCheck.ml code - but I think it's the
lesser of evils for now because we don't want to expose two different
interfaces with the same name, and at the moment I can't come up with two
good names (partly because I don't understand the behavior fully,
which typically makes it hard to come up with great names!).

Reviewed By: grievejia

Differential Revision: D52101099

fbshipit-source-id: 862342c7150db1d62bb563d19a85c0031effe6e3