Make attribute_from_class_name logic-only
Summary:
In this diff I continue to split the GlobalResolution functionality
into simple proxy functions that grab underlying environments and
query them (these eventually will be callbacks in a record) versus
logical functions that use the proxies to add additional layers of
functionality.

In this case, I make `attribute_from_class_name` a logical function,
by extracting the `attribute` function as a proxy (private, it's not
exposed in the interface since nothing downstream of GlobalResolution
needs it) and using the preexisting `get_class_summary` proxy instead
of using UGE's get_class_summary directly.

Reviewed By: grievejia

Differential Revision: D52101093

fbshipit-source-id: 31000918cbe8f2f08926d03cf37bb5843df851c7