Move `global` logic
Summary:
This diff begins a stack to refactor the internals of GlobalResolution until
I can easily use an abstract API to mediate access to lower environment
layers rather than directly wrapping them. (The previous stack / stack up till this
point was to refactor the *interface* and callsites, the shift here is in focus
toward internals rather than interface).

The first thing I need to do is split GlobalResolution's functions into
two kinds:
- proxied functions that just call underlying environments. These will
  eventually become functions in the record.
- functions that implement additional functionality *in terms of* the
  underlying environment. These will remain unchanged.

To kick off this effort, I'm pushing the definition of `global`, which is
a thin wrapper around `AttributeResolution` logic already to intercept a
few special cases, down into `AttributeResolution`.

This requires creating an OutgoingDataComputation module there, since we
want to keep the compute and storage separated so we don't want to inline
the logic into `AttributeResolution.ReadOnly`.

Reviewed By: grievejia

Differential Revision: D52101096

fbshipit-source-id: da0c721a455fa725b2048a9bc433380a0117b856