Reorganize globalResolution
Summary:
This is pure moving things around.

Isolating all proxied functions at the top, in order of
environment layer. Trying to roughly group logical functions
by associated environment layer as well.

In addition, making the .ml and .mli files agree on the order
to make navigating the two of them a little easier.

Reviewed By: grievejia

Differential Revision: D52101094

fbshipit-source-id: ae7f78c8857353889fcc671907d07131fddd84f9