Allow file updates from files not tracked by build system
Summary:
The build system keeps a mapping of artifact path <> source path. This mapping is necessary when a build system has its own location for a file.

There are currently cases where we are using a file not in the build system (e.g. non-buck folders within a buck repository - see D49383381). When there are file updates here, we should revert to the default artifact path (the file path using python semantics).

We don't need to worry about excluding files here since every file that the LS asks for can be analyzed by the code navigation server.

Reviewed By: grievejia

Differential Revision: D52181816

fbshipit-source-id: dc27cd0eeb4ef900e62e6d4f179f304b9c3c0172